i-raw: bring in line with dnglab git
remember to run `cargo update` in the `i-raw/rawloader-c/` subfolder for old checkouts.

+++ src/pipe/modules/i-raw/rawloader-c/rawloader.h
@@ -11,8 +11,8 @@ typedef struct rawimage_t
   uint64_t height;
   uint64_t cpp;
   float    wb_coeffs[4];
-  uint16_t whitelevels[4];
-  uint16_t blacklevels[4];
+  float    whitelevels[4];
+  float    blacklevels[4];
   float    xyz_to_cam[4][3];
   uint32_t filters;
   uint64_t crop_aabb[4];
